+++ Cargo.toml
@@ -30,6 +30,7 @@ members = [
   "ext/websocket",
   "ext/webstorage",
   "ext/napi",
+  "lockfile",
 ]
 exclude = ["test_util/std/hash/_wasm"]
 
@@ -50,6 +51,7 @@ deno_runtime = { version = "0.93.0", path = "./runtime" }
 napi_sym = { version = "0.15.0", path = "./cli/napi/sym" }
 deno_bench_util = { version = "0.79.0", path = "./bench_util" }
 test_util = { path = "./test_util" }
+deno_lockfile = { version = "0.1.0", path = "./lockfile" }
 
 # exts
 deno_broadcast_channel = { version = "0.79.0", path = "./ext/broadcast_channel" }
