+++ cli/Cargo.toml
@@ -48,6 +48,7 @@ deno_doc = "0.52.0"
 deno_emit = "0.13.0"
 deno_graph = "0.41.0"
 deno_lint = { version = "0.37.0", features = ["docs"] }
+deno_lockfile.workspace = true
 deno_runtime.workspace = true
 deno_task_shell = "0.8.1"
 napi_sym.workspace = true
