+++ cli/args/mod.rs
@@ -502,7 +502,7 @@ impl CliOptions {
   pub fn from_flags(flags: Flags) -> Result<Self, AnyError> {
     let maybe_config_file = ConfigFile::discover(&flags)?;
     let maybe_lock_file =
-      Lockfile::discover(&flags, maybe_config_file.as_ref())?;
+      lockfile::discover(&flags, maybe_config_file.as_ref())?;
     Ok(Self::new(flags, maybe_config_file, maybe_lock_file))
   }
